ai-azad/backend/requirements.txt
flask==2.0.3
pymongo==4.0.1
psycopg2-binary==2.9.3
elasticsearch==7.10.0
redis==4.0.2
transformers==4.19.2
torch==1.10.2
flask-cors==3.0.10
flask-jwt-extended==4.3.1
pytesseract==0.3.8
pdf2image==1.16.3
pytest==6.2.5